requirements.txt
PyQt5
pyqtgraph
numpy
matplotlib
opencv-python
PyOpenGL
qimage2ndarray
vedo
pydantic
pillow
